test: split test-cli-node-options
* partitioning the subprocess groups
* use `-e` instead of module
* reduce maxBuffer

PR-URL: https://github.com/nodejs/node/pull/14195
Fixes: https://github.com/nodejs/node/issues/14191
